Fix move state and zoom gui operations.

+++ utils.cpp
@@ -3,7 +3,7 @@
 #include "utils.h"
 
 
-bool point_in_box(int pt_x, int pt_y, int box_x, int box_y, int box_w, int box_h)
+bool point_in_box(float pt_x, float pt_y, float box_x, float box_y, float box_w, float box_h)
 {
     return pt_x >= box_x && pt_x <= box_x + box_w &&
         pt_y >= box_y && pt_y <= box_y + box_h;
