+++ utils.h
@@ -5,7 +5,7 @@
 using std::string;
 
 
-bool point_in_box(int pt_x, int pt_y, int box_x, int box_y, int box_w, int box_h);
+bool point_in_box(float pt_x, float pt_y, float box_x, float box_y, float box_w, float box_h);
 
 
 class IOResult
